feat: integrate geoEventsService
